feat: add public landing page at /, move project list to /projects
The root URL was the gated project list, so a logged-out visitor saw a
blank page and a redirect to /login with no explanation of what Catalyse
is or why to join.

- / is now a public landing page explaining PauseAI UK, the kinds of work
  volunteers do, and how the platform works. It fetches no project data,
  so nothing about specific projects leaks to unauthenticated visitors.
- The project list moves to /projects, unchanged and still gated behind
  useRequireApproved.
- The header wordmark points at /projects for signed-in volunteers and /
  for visitors, preserving the existing "logo goes to the app" behaviour.
- Redirect targets that meant "the project list" (admin guards, project
  not-found, post-delete) now point at /projects rather than /.

Content is drawn from pauseai.uk and pauseai.info. No statistics or
membership numbers are claimed, so the copy does not go stale.

Closes #106
