+++ app/page.tsx
@@ -1,390 +1,209 @@
-'use client'
-
-import { useEffect, useState } from 'react'
-import { useRequireApproved } from '@/lib/hooks/auth'
+import type { Metadata } from 'next'
 import Link from 'next/link'
-import { useQuery } from '@tanstack/react-query'
-import Button from '@/components/Button'
-import FilterDropdown, { useFilterOptions } from '@/components/FilterDropdown'
-import { buildLocationOptions, type LocalGroupOption } from '@/lib/filter-options'
-import { InferRouterInputs } from '@orpc/server'
-import { ORPCError } from '@orpc/client'
-import { orpc } from '@/lib/orpc'
-import { AppRouter } from '@/server/router'
-import { type Project, ProjectList, statusBadgeClasses } from '@/components/ProjectCard'
-import { ProjectStatus } from '@/generated/prisma/enums'
+import LandingCTA from '@/components/LandingCTA'
 
-const STATUS_OPTIONS = [
-  { value: '', label: 'All Active' },
-  { value: 'in_progress', label: 'In Progress' },
-  { value: 'on_hold', label: 'On Hold' },
-  { value: 'completed', label: 'Completed' },
-  { value: 'archived', label: 'Archived' },
-] as const
+export const metadata: Metadata = {
+  title: 'Volunteer with PauseAI UK',
+  description:
+    'Catalyse is the volunteer platform for PauseAI UK — find campaign, policy, organising and creative work that matches your skills.',
+}
 
-const NEEDS_OPTIONS = [
-  { value: '', label: 'Any' },
-  { value: 'looking_for_people', label: 'Looking for People' },
-  { value: 'seeking_help', label: 'Seeking Help', indent: true },
-  { value: 'seeking_owner', label: 'Seeking Owner', indent: true },
-  { value: 'not_seeking', label: 'Not Seeking' },
-] as const
+/**
+ * Public landing page.
+ *
+ * Deliberately contains no project data. Individual projects are only visible
+ * to approved volunteers at /projects; this page describes the *kinds* of work
+ * the movement does so a prospective volunteer can decide whether to apply.
+ */
 
-const URGENCY_OPTIONS = [
-  { value: '', label: 'Any urgency' },
-  { value: 'high', label: 'High' },
-  { value: 'medium', label: 'Medium' },
-  { value: 'low', label: 'Low' },
-] as const
+const WORK_AREAS: { title: string; body: string }[] = [
+  {
+    title: 'Campaigns & protests',
+    body: 'Organising demonstrations, planning actions, stewarding on the day, and the logistics that hold a public campaign together.',
+  },
+  {
+    title: 'Political engagement',
+    body: 'Writing to MPs, preparing briefings for lawmakers, drafting policy proposals, and following up after meetings.',
+  },
+  {
+    title: 'Local chapters',
+    body: 'Running regular meetups, welcoming new volunteers, and starting a chapter where there isn’t one yet.',
+  },
+  {
+    title: 'Events',
+    body: 'Conferences, workshops, film screenings, book launches and socials — from finding a venue to running the door.',
+  },
+  {
+    title: 'Communications & media',
+    body: 'Social media, newsletters, press outreach, and helping people tell their own stories about why this matters to them.',
+  },
+  {
+    title: 'Design & creative',
+    body: 'Graphics, placards, video, photography, merchandise, and the visual identity that makes a campaign recognisable.',
+  },
+  {
+    title: 'Research & writing',
+    body: 'Explainers, fact-checking, literature summaries, and turning dense technical material into something a non-specialist can act on.',
+  },
+  {
+    title: 'Software & operations',
+    body: 'Internal tools, data and analysis, web work, and the unglamorous admin that keeps a volunteer organisation running.',
+  },
+]
 
-const SORT_OPTIONS = [
-  { value: '', label: 'Default sort' },
-  { value: 'created_at', label: 'Newest first' },
-  { value: 'match', label: 'Best match' },
-  { value: 'urgency', label: 'Most urgent' },
-] as const
+const STEPS: { title: string; body: string }[] = [
+  {
+    title: 'Apply',
+    body: 'Tell us about your connection to PauseAI, what you’re good at, and how much time you have. Applications are read by a person, not a filter.',
+  },
+  {
+    title: 'Get matched',
+    body: 'Once you’re approved you can see every open project. We highlight the ones that need the skills you listed, so you don’t have to guess where you’re useful.',
+  },
+  {
+    title: 'Contribute',
+    body: 'Take on a single task, join a project team, or lead a project of your own. If you have an idea we aren’t working on yet, you can propose it.',
+  },
+]
 
-export default function ProjectsPage() {
-  const { user, loading } = useRequireApproved()
-  const userSkillIds = new Set(user?.skills?.map((s) => s.id) ?? [])
-  const [search, setSearch] = useState('')
-  const { value: statusFilter, onChange: setStatusFilter } = useFilterOptions(STATUS_OPTIONS, '')
-  const { value: needsFilter, onChange: setNeedsFilter } = useFilterOptions(NEEDS_OPTIONS, '')
-  const { value: urgencyFilter, onChange: setUrgencyFilter } = useFilterOptions(URGENCY_OPTIONS, '')
-  const [locationFilter, setLocationFilter] = useState('')
-  const { value: sortBy, onChange: setSortBy } = useFilterOptions(SORT_OPTIONS, '')
-  const [completedOpen, setCompletedOpen] = useState(false)
-  const [debouncedSearch, setDebouncedSearch] = useState('')
-  const [debouncedLocationFilter, setDebouncedLocationFilter] = useState('')
-
-  useEffect(() => {
-    const t = setTimeout(
-      () => {
-        setDebouncedSearch(search)
-        setDebouncedLocationFilter(locationFilter)
-      },
-      search || locationFilter ? 300 : 0,
-    )
-    return () => clearTimeout(t)
-  }, [search, locationFilter])
-
-  const { data: pendingTriageList = [] } = useQuery({
-    ...orpc.admin.triage.list.queryOptions(),
-    enabled: !!user?.isAdmin,
-  })
-  const pendingCount = pendingTriageList.length
-
-  const { data: pendingApplicationsList = [] } = useQuery({
-    ...orpc.admin.applications.list.queryOptions({ input: { filter: 'mine' } }),
-    enabled: !!user?.isAdmin,
-  })
-  const pendingApplicationsCount = pendingApplicationsList.length
-
-  const { data: localGroupsData } = useQuery({
-    ...orpc.localGroups.list.queryOptions({ input: {} }),
-    enabled: true,
-  })
-  const localGroups: LocalGroupOption[] = localGroupsData?.groups ?? []
-
-  const projectsInput: InferRouterInputs<AppRouter>['projects']['list'] = {}
-  if (debouncedSearch) projectsInput.search = debouncedSearch
-  if (statusFilter) projectsInput.status = statusFilter
-  if (needsFilter === 'looking_for_people') projectsInput.isSeekingAny = true
-  else if (needsFilter === 'seeking_help') projectsInput.isSeekingHelp = true
-  else if (needsFilter === 'seeking_owner') projectsInput.isSeekingOwner = true
-  else if (needsFilter === 'not_seeking') projectsInput.notSeeking = true
-  if (urgencyFilter) projectsInput.urgency = urgencyFilter
-  if (debouncedLocationFilter) {
-    const [country, localGroup] = debouncedLocationFilter.split(':')
-    projectsInput.country = country
-    if (localGroup) projectsInput.localGroup = localGroup
-  }
-  if (sortBy) projectsInput.sortBy = sortBy
-
-  const {
-    data: projectsData,
-    isPending: loadingProjects,
-    error: projectsError,
-  } = useQuery({
-    ...orpc.projects.list.queryOptions({ input: projectsInput }),
-    enabled: !!user,
-  })
-  const projects = projectsData?.projects ?? []
-
-  const hasFilters =
-    search || statusFilter || needsFilter || urgencyFilter || locationFilter || sortBy
-
-  function clearFilters() {
-    setSearch('')
-    setStatusFilter('')
-    setNeedsFilter('')
-    setUrgencyFilter('')
-    setLocationFilter('')
-    setSortBy('')
-  }
-
-  function byMatchScore(a: Project, b: Project) {
-    const scoreA = a.match?.matchedRequiredCount ?? 0
-    const scoreB = b.match?.matchedRequiredCount ?? 0
-    return scoreB - scoreA
-  }
-  const sortGroup = (list: Project[]) =>
-    userSkillIds.size > 0 ? [...list].sort(byMatchScore) : list
-
-  const seeking = sortGroup(projects.filter((p) => p.isSeekingHelp || p.isSeekingOwner))
-  const inProgress = sortGroup(
-    projects.filter(
-      (p) => !p.isSeekingHelp && !p.isSeekingOwner && p.status === ProjectStatus.in_progress,
-    ),
-  )
-  const onHold = sortGroup(
-    projects.filter(
-      (p) => !p.isSeekingHelp && !p.isSeekingOwner && p.status === ProjectStatus.on_hold,
-    ),
-  )
-  const completed = sortGroup(projects.filter((p) => p.status === ProjectStatus.completed))
-  const other = sortGroup(
-    projects.filter(
-      (p) =>
-        !p.isSeekingHelp &&
-        !p.isSeekingOwner &&
-        !['in_progress', 'on_hold', 'completed'].includes(p.status),
-    ),
-  )
-
-  const groups = [
-    {
-      key: 'seeking',
-      projects: seeking,
-      label: 'Looking for People',
-      desc: 'These projects need your help',
-      color: 'text-orange-600 dark:text-orange-400',
-    },
-    {
-      key: 'in_progress',
-      projects: inProgress,
-      label: 'In Progress',
-      desc: 'Actively being worked on',
-      color: 'text-blue-600 dark:text-blue-400',
-    },
-    { key: 'other', projects: other, label: 'Other Active', desc: '', color: 'text-text-light' },
-    {
-      key: 'on_hold',
-      projects: onHold,
-      label: 'On Hold',
-      desc: '',
-      color: 'text-red-600 dark:text-red-400',
-    },
-    {
-      key: 'completed',
-      projects: completed,
-      label: 'Completed',
-      desc: '',
-      color: 'text-green-600 dark:text-green-400',
-    },
-  ]
-
-  if (loading || !user) return null
-
+export default function LandingPage() {
   return (
-    <>
-      <main className="container py-5 pb-15">
-        <h1 role="heading">Projects</h1>
-
-        {user.isAdmin && pendingCount > 0 && (
-          <div className="flex items-center gap-3 p-4 rounded-lg mb-4 bg-amber-100 text-amber-800 border border-amber-300 dark:bg-amber-900 dark:text-amber-200 dark:border-amber-600">
-            <strong>
-              {pendingCount} project{pendingCount !== 1 ? 's' : ''} pending review.
-            </strong>{' '}
-            <Link href="/admin/triage" className="underline">
-              Go to triage →
-            </Link>
-          </div>
-        )}
-        {user.isAdmin && pendingApplicationsCount > 0 && (
-          <div className="flex items-center gap-3 p-4 rounded-lg mb-4 bg-violet-100 text-violet-800 border border-violet-300 dark:bg-violet-950 dark:text-violet-300 dark:border-violet-600">
-            <strong>
-              {pendingApplicationsCount} application{pendingApplicationsCount !== 1 ? 's' : ''}{' '}
-              pending review.
-            </strong>{' '}
-            <Link href="/admin/applications" className="underline">
-              Review applications →
-            </Link>
-          </div>
-        )}
-
-        {/* Filters */}
-        <div className="mb-5">
-          <div className="mb-3">
-            <label htmlFor="search-projects">Search</label>
-            <input
-              id="search-projects"
-              type="search"
-              aria-label="Search"
-              placeholder="Search projects…"
-              value={search}
-              onChange={(e) => setSearch(e.target.value)}
-            />
-          </div>
-          <div className="flex gap-3 flex-wrap items-end">
-            <FilterDropdown
-              id="status-filter"
-              label="Status"
-              ariaLabel="Status filter"
-              value={statusFilter}
-              options={STATUS_OPTIONS}
-              onChange={setStatusFilter}
-            />
-            <FilterDropdown
-              id="needs-filter"
-              label="Needs"
-              ariaLabel="Needs filter"
-              value={needsFilter}
-              options={NEEDS_OPTIONS}
-              onChange={setNeedsFilter}
-            />
-            <FilterDropdown
-              id="urgency-filter"
-              label="Urgency"
-              ariaLabel="Urgency filter"
-              value={urgencyFilter}
-              options={URGENCY_OPTIONS}
-              onChange={setUrgencyFilter}
-            />
-
-            <FilterDropdown
-              id="location-filter"
-              label="Country/Group"
-              ariaLabel="Country/Group filter"
-              value={locationFilter}
-              options={buildLocationOptions(localGroups)}
-              onChange={setLocationFilter}
-              searchable
-            />
-
-            <FilterDropdown
-              id="sort-filter"
-              label="Sort by"
-              ariaLabel="Sort filter"
-              value={sortBy}
-              options={SORT_OPTIONS}
-              onChange={setSortBy}
-            />
-
-            {hasFilters && (
-              <Button variant="outline" size="lg" onClick={clearFilters}>
-                Clear filters
-              </Button>
-            )}
+    <main>
+      {/* Hero */}
+      <section className="border-b border-brand-border bg-surface">
+        <div className="container py-14 md:py-20">
+          <div className="max-w-3xl mx-auto md:text-center">
+            <p className="font-heading text-sm font-bold uppercase tracking-widest text-primary-dark mb-3">
+              PauseAI UK
+            </p>
+            <h1 className="text-3xl md:text-5xl">Find the work that needs you</h1>
+            <p className="text-lg text-text-light mb-8">
+              Catalyse is the volunteer platform for PauseAI UK. It connects the people who want to
+              help with the projects that need them — matching what you can do to what the movement
+              is actually working on right now.
+            </p>
+            <div className="md:flex md:justify-center">
+              <LandingCTA />
+            </div>
+            <p className="text-sm text-text-light mt-4 mb-0">
+              Already applied? <Link href="/login">Log in to check your application status</Link>.
+            </p>
           </div>
         </div>
+      </section>
 
-        {loadingProjects ? (
-          <div className="text-center py-10 text-text-light">Loading projects…</div>
-        ) : projectsError ? (
-          <div className="text-center py-15 px-5 text-text-light">
-            <h3>Couldn’t load projects</h3>
-            <p>
-              {projectsError instanceof Error
-                ? projectsError.message
-                : 'Something went wrong loading projects.'}
+      {/* What PauseAI UK is */}
+      <section className="container py-12 md:py-16">
+        <div className="max-w-3xl">
+          <h2>Who we are</h2>
+          <p className="text-text-light">
+            PauseAI UK is a civic movement working to avert the risks of superhuman artificial
+            intelligence. We campaign for binding limits on the development of the most powerful AI
+            systems until there is a credible way to make them safe — through public campaigning,
+            political engagement, and local organising across the UK.
+          </p>
+          <p className="text-text-light">
+            We are a small staff supported by a large number of volunteers, with chapters in cities
+            including London, Glasgow, Oxford, Leicester, Manchester and the West of England. You do
+            not need a technical background, and you do not need to be an expert on AI.
+          </p>
+          <p className="mb-0">
+            <a href="https://pauseai.uk" target="_blank" rel="noopener noreferrer">
+              Read more about PauseAI UK →
+            </a>
+          </p>
+        </div>
+      </section>
+
+      {/* Kinds of work */}
+      <section className="bg-surface border-y border-brand-border">
+        <div className="container py-12 md:py-16">
+          <div className="max-w-3xl mb-8">
+            <h2>The kinds of things we do</h2>
+            <p className="text-text-light mb-0">
+              Projects on Catalyse change constantly, but they tend to fall into these areas. Some
+              need a few hours once; others need someone to take ownership for months.
             </p>
-            {projectsError instanceof ORPCError && projectsError.code === 'FORBIDDEN' && (
-              <p className="mt-2">
-                <Link href="/verify-email" className="underline">
-                  Confirm your email
-                </Link>
-              </p>
-            )}
           </div>
-        ) : projects.length === 0 ? (
-          <div className="text-center py-15 px-5 text-text-light">
-            <h3>No projects found</h3>
-            <p>
-              Try adjusting your filters or{' '}
-              <Link href="/suggest" className="underline">
-                suggest a new project
-              </Link>
-              .
-            </p>
-          </div>
-        ) : (
-          <>
-            {/* Status summary bar */}
-            {!statusFilter && !needsFilter && projects.length > 1 && (
-              <div className="flex flex-wrap gap-2 mb-6">
-                {seeking.length > 0 && (
-                  <span className={statusBadgeClasses('seeking_help')}>
-                    Looking for People: {seeking.length}
-                  </span>
-                )}
-                {inProgress.length > 0 && (
-                  <span className={statusBadgeClasses('in_progress')}>
-                    In Progress: {inProgress.length}
-                  </span>
-                )}
-                <span className="text-sm text-text-light self-center">{projects.length} total</span>
+          <div className="grid grid-cols-1 md:grid-cols-2 xl:grid-cols-4 gap-5">
+            {WORK_AREAS.map((area) => (
+              <div
+                key={area.title}
+                className="bg-brand-bg border border-brand-border rounded-xl p-5"
+              >
+                <h3 className="text-base mb-2">{area.title}</h3>
+                <p className="text-sm text-text-light mb-0">{area.body}</p>
               </div>
-            )}
+            ))}
+          </div>
+        </div>
+      </section>
 
-            {/* Grouped project cards */}
-            {statusFilter || needsFilter ? (
-              <ProjectList projects={projects} userSkillIds={userSkillIds} />
-            ) : (
-              groups
-                .filter((g) => g.projects.length > 0)
-                .map((g) => {
-                  const isCompleted = g.key === 'completed'
-                  const isOpen = !isCompleted || completedOpen
-                  return (
-                    <div key={g.key} className="mb-8">
-                      {isCompleted ? (
-                        <h2
-                          className="text-lg mb-1 flex items-center gap-2 cursor-pointer select-none"
-                          onClick={() => setCompletedOpen((o) => !o)}
-                          role="button"
-                          aria-expanded={completedOpen}
-                          tabIndex={0}
-                          onKeyDown={(e) => e.key === 'Enter' && setCompletedOpen((o) => !o)}
-                        >
-                          {g.label} — {g.projects.length} project
-                          {g.projects.length !== 1 ? 's' : ''}
-                          <svg
-                            className={`text-text-light shrink-0 transition-transform ${completedOpen ? 'rotate-180' : 'rotate-0'}`}
-                            width="32"
-                            height="32"
-                            viewBox="0 0 24 24"
-                            fill="none"
-                            stroke="currentColor"
-                            strokeWidth="2.5"
-                            strokeLinecap="round"
-                            strokeLinejoin="round"
-                          >
-                            <polyline points="6 9 12 15 18 9" />
-                          </svg>
-                        </h2>
-                      ) : (
-                        <h2 className={`text-lg mb-1 ${g.color}`}>
-                          {g.label} — {g.projects.length} project
-                          {g.projects.length !== 1 ? 's' : ''}
-                        </h2>
-                      )}
-                      {g.desc && <p className="text-text-light text-sm mb-3">{g.desc}</p>}
-                      {isOpen && (
-                        <div
-                          key={String(completedOpen)}
-                          className={isCompleted ? 'animate-fade-slide-in' : undefined}
-                        >
-                          <ProjectList projects={g.projects} userSkillIds={userSkillIds} />
-                        </div>
-                      )}
-                    </div>
-                  )
-                })
-            )}
-          </>
-        )}
-      </main>
-    </>
+      {/* How it works */}
+      <section className="container py-12 md:py-16">
+        <div className="max-w-3xl mb-8">
+          <h2>How Catalyse works</h2>
+        </div>
+        <ol className="grid grid-cols-1 md:grid-cols-3 gap-5 list-none p-0 m-0">
+          {STEPS.map((step, i) => (
+            <li key={step.title} className="bg-surface border border-brand-border rounded-xl p-6">
+              <span
+                className="inline-flex items-center justify-center w-8 h-8 rounded-full bg-primary text-[#111827] font-bold mb-3"
+                aria-hidden="true"
+              >
+                {i + 1}
+              </span>
+              <h3 className="text-lg mb-2">
+                <span className="sr-only">Step {i + 1}: </span>
+                {step.title}
+              </h3>
+              <p className="text-sm text-text-light mb-0">{step.body}</p>
+            </li>
+          ))}
+        </ol>
+      </section>
+
+      {/* Expectations */}
+      <section className="bg-surface border-y border-brand-border">
+        <div className="container py-12 md:py-16">
+          <div className="max-w-3xl">
+            <h2>Before you apply</h2>
+            <ul className="text-text-light space-y-2 pl-5 list-disc marker:text-primary">
+              <li>
+                Applications are reviewed by our team, so there is a short wait before you can see
+                open projects. We will email you either way.
+              </li>
+              <li>
+                There is no minimum commitment. Plenty of volunteers contribute an hour here and
+                there, and that is genuinely useful.
+              </li>
+              <li>
+                You control what other volunteers can see. Your profile can stay out of the
+                directory entirely, and your contact details are never shared without your say-so.
+              </li>
+              <li>
+                Catalyse is for coordinating work. For events, the newsletter and the wider
+                community, start at{' '}
+                <a href="https://pauseai.uk" target="_blank" rel="noopener noreferrer">
+                  pauseai.uk
+                </a>
+                .
+              </li>
+            </ul>
+          </div>
+        </div>
+      </section>
+
+      {/* Closing CTA */}
+      <section className="container py-14 md:py-20">
+        <div className="max-w-2xl">
+          <h2>Ready to start?</h2>
+          <p className="text-text-light mb-6">
+            Tell us what you are good at and we will show you where it is needed.
+          </p>
+          <LandingCTA />
+        </div>
+      </section>
+    </main>
   )
 }
