+++ app/projects/layout.tsx
@@ -1,0 +1,7 @@
+import type { Metadata } from 'next'
+
+export const metadata: Metadata = { title: 'All Projects' }
+
+export default function Layout({ children }: { children: React.ReactNode }) {
+  return <>{children}</>
+}
